Moved from outdated and unmaintained cwautomacros to the cwm4 git submodule.
The downside is that that kinda assumes it is a C++ project,
but if that is a serious problem then I can change cwm4.
All it really does is look for a C++ compiler, so a developer
needs to have one installed (it won't be used obviously).

(cherry picked from commit 8c49839474410afa735d4499bc69da7472fce1f7)