Add submodule `indent` back.
This submodule is only needed for the "maintainer", since it contains
`man/texinfo2man.c` that is used to generate the man page (which.1).

`indent` is not added to the distribution tarball.

Note that everyone that checks out `which` using git needs to configure
using --enable-maintainer-mode (after running ./autogen.sh; which in
turn requires cwautomacros to be *installed*; see
https://github.com/CarloWood/cwautomacros).