Add a README.md now that we're on git.
We're not on github though, so not sure this is going to work.
This README was written by Mitch Capper, see
https://github.com/mitchcapper/which/commit/7ab5cb6b05db1db1fcadac22bad0b3585511ec0b

However, that version contained carriage return characters.
Also, I don't think a version jump from 2.21 to 2.40 is warranted (at
all).

I did not include the "2.4 added better windows support." the history,
that belongs in NEWS.