Fix the bug introduced by a9a25e49c0c0db45468d7af0e407569518b4e2b5
Without this fix you'd find the same executables
with "ls" as with "./ls":

$ which ./ls
/usr/bin/ls

while the correct reply is:
$ which ./ls
/usr/bin/which: no ls in (.)

Having any '/' at all in the search name means that
you DO specify a directory, either absolute, or relative
to the current directory (./...) or relative to your
home direction (~/...) and in all cases the PATH should
no longer be searched.

As the naming in the code (is_absolute) was just wrong, I can
understand this confusion; therefore I renamed the function
and even local variables as well to something that more clearly
states that this is about having a directory separator in the
name, NOT if it is an absolute path.